refactor: extract URL-handler bundle build into pkg/urihandler
Move the bundle build/teardown out of pkg/commands/accesshandler/actions
into a feature library so cobra wrappers stay thin and the logic is callable
from non-cobra code paths.

- Move register/unregister business logic to pkg/urihandler/
- Extract handler.sh and handler.applescript to dedicated files, embedded
  via //go:embed
- Slim pkg/commands/accesshandler/actions/{register,unregister}.go to
  cobra wrappers
