+++ pkg/commands/accesshandler/actions/unregister.go
@@ -1,13 +1,9 @@
 package actions
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"os/exec"
-	"runtime"
-
 	"github.com/spf13/cobra"
+
+	"github.com/apono-io/apono-cli/pkg/urihandler"
 )
 
 func Unregister() *cobra.Command {
@@ -15,34 +11,7 @@ func Unregister() *cobra.Command {
 		Use:   "unregister",
 		Short: "Remove the apono:// URL handler",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if runtime.GOOS != darwinOS {
-				return fmt.Errorf("access-handler is only supported on macOS")
-			}
-
-			bundleDir, err := bundlePath()
-			if err != nil {
-				return err
-			}
-
-			if _, statErr := os.Stat(bundleDir); os.IsNotExist(statErr) {
-				_, err = fmt.Fprintln(cmd.OutOrStdout(), "Protocol handler is not registered")
-				return err
-			}
-
-			if out, lsErr := unregisterFromLaunchServices(bundleDir); lsErr != nil {
-				return fmt.Errorf("lsregister -u: %w: %s", lsErr, string(out))
-			}
-
-			if err = os.RemoveAll(bundleDir); err != nil {
-				return fmt.Errorf("remove %s: %w", bundleDir, err)
-			}
-
-			_, err = fmt.Fprintln(cmd.OutOrStdout(), "Unregistered apono:// handler")
-			return err
+			return urihandler.Unregister(cmd.OutOrStdout())
 		},
 	}
-}
-
-func unregisterFromLaunchServices(bundleDir string) ([]byte, error) {
-	return exec.CommandContext(context.Background(), lsregisterPath, "-u", bundleDir).CombinedOutput()
 }
